Interactive Command Line functions!
After data load the user can now run functions to aggregate some data
about a specifc stock on a certain day. I am not sure if this is how
we are supposed to do it but I was quite bored on the train and I hacked
this out. If we want the data shoved in the DB I guess we could do that too
but that would mean doing some real normalization on the data. A huge chunk
of it is repeated A LOT, so we would have to do some work... :dolphin: