Should fix Month-Day confusion
Prior to this commit the expression used to do a date based search
could return data from a day with a matchin number (i.e. 02 is both a
viable day and month number). This now checks the length of the incoming
"date" string and modifies the query appropriatley. This seems to be
passing the totally rigorous tests I totally put it through. :tada: