Moves Input loop to own method, cleans up close logic
The loop to get input and run functions is now in its own method for
cleanliness's sake. Also moved the Connection.close() call to only
occur if the Connection is not null. I decided to do that after
noticing a null pointer error that would occur if the Connection
did not open, the program would still try and close it. :hammer: :bug: